requirements.txt
hikari[speedups]==2.0.0.dev117
hikari-tanjun==2.12.0
hikari-miru==3.0.2
pyyaml==6.0.0
ormar[orjson,postgresql]==0.12.1
uvloop==0.17; platform_system != "Windows"
